requirements.txt
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
google-generativeai>=0.8.0
playwright>=1.40.0
modelcontextprotocol>=1.0.1
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0
openai>=1.12.0
anthropic>=0.21.0
schedule>=1.2.0
requests>=2.31.0
pytest>=7.4.0
python-multipart>=0.0.6
duckduckgo-search>=6.0.0
websockets>=12.0
